+++ lib/agilibox/cucumber_helpers/capybara_selectors.rb
@@ -1,8 +1,7 @@
 Capybara.add_selector(:agilibox_clickable) do
   xpath do |locator, **options|
-    self.class.all
-      .values_at(:link, :button, :label)
-      .map { |selector| instance_exec(locator, **options, &selector.xpath) }
+    %i(link button label)
+      .map { |selector| expression_for(selector, locator, **options) }
       .reduce(:union)
   end
 end